+++ src/libclient/export/animationsaverrunnable.h
@@ -20,7 +20,7 @@ public:
 		const QString &path,
 #endif
 		int format, int width, int height, int loops, int start, int end,
-		int framerate, const QRect &crop, bool scaleSmooth,
+		double framerate, const QRect &crop, bool scaleSmooth,
 		const drawdance::CanvasState &canvasState, QObject *parent = nullptr);
 
 	void run() override;
@@ -57,7 +57,7 @@ private:
 	const int m_loops;
 	const int m_start;
 	const int m_end;
-	const int m_framerate;
+	const double m_framerate;
 	const QRect m_crop;
 	const drawdance::CanvasState m_canvasState;
 	const bool m_scaleSmooth;
